Asset Tracker: finalize SPI1 stress test application.
