+++ user/tests/wiring/spi1_stress/esp32_soc.h
